changed response model from int to long

+++ Plugins/BTCPayServer.Plugins.NairaCheckout/ViewModels/MavapayVM/CreateQuoteResponseVm.cs
@@ -5,12 +5,12 @@ namespace BTCPayServer.Plugins.NairaCheckout.ViewModels;
 public class CreateQuoteResponseVm
 {
     public string id { get; set; }
-    public int exchangeRate { get; set; }
+    public long exchangeRate { get; set; }
     public double usdToTargetCurrencyRate { get; set; }
     public string sourceCurrency { get; set; }
     public string targetCurrency { get; set; }
-    public int transactionFeesInSourceCurrency { get; set; }
-    public int transactionFeesInTargetCurrency { get; set; }
+    public long transactionFeesInSourceCurrency { get; set; }
+    public long transactionFeesInTargetCurrency { get; set; }
     public long amountInSourceCurrency { get; set; }
     public long amountInTargetCurrency { get; set; }
     public string paymentMethod { get; set; }
@@ -18,12 +18,12 @@ public class CreateQuoteResponseVm
     public bool isValid { get; set; }
     public string invoice { get; set; }
     public string hash { get; set; }
-    public int totalAmountInSourceCurrency { get; set; }
-    public int customerInternalFee { get; set; }
+    public long totalAmountInSourceCurrency { get; set; }
+    public long customerInternalFee { get; set; }
     public string bankName { get; set; }
     public string ngnBankAccountNumber { get; set; }
     public string ngnAccountName { get; set; }
-    public int estimatedRoutingFee { get; set; }
+    public long estimatedRoutingFee { get; set; }
     public string orderId { get; set; }
 }
